Added Join Commands
Join commands are something people are wanting (well, my client base),
so heres a small work. It allows currently the following:

http://i.imgur.com/BQULKVm.png

It does not use a class at this time, just a dictionary (I never tested
if the custom class was supported if it was in both the server, and
client project).